# swagger.full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PIM API",
    "version": "1.4.1"
  },
  "paths": {
...
    "/api/transaction/csv": {
      "post": {
        "description": "Creates new products from CSV file",
        "parameters": [
          {
            "name": "template",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "CSV template (sryhma, kesko, default)"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactions": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Products added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/transaction": {
      "post": {
        "description": "Creates a new transaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/receipt/original": {
      "post": {
        "description": "Uploads an original receipt file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "src": {
                    "type": "string",
                    "format": "binary"
                  },
                  "id": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Receipt uploaded",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/receipt/edit": {
      "post": {
        "description": "Uploads and crops a receipt file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "src": {
                    "type": "string",
                    "format": "binary"
                  },
                  "id": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Receipt uploaded",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/receipt/data/original/{id}": {
      "post": {
        "description": "Creates receipt from original receipt file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Receipt ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/receipt/data/edited/{id}": {
      "post": {
        "description": "Creates receipt from edited receipt file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Receipt ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "contributionList": {
            "type": "string",
            "description": "Product contributions separated with comma"
          },
          "measure": {
            "type": "number",
            "description": "Product amount measure"
          },
          "unit": {
            "type": "string",
            "description": "Product amount unit"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductAttribute"
            },
            "description": "List of product attributes"
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductContribution"
            }
          }
        }
      },
      "ProductContribution": {
        "type": "object",
        "required": ["id", "amount", "unit"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "contribution": {
            "$ref": "#/components/schemas/Category"
          }
        }
      },
      "Attribute": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "object",
            "properties": {
              "fi-FI": {
                "type": "string"
              },
              "en-US": {
                "type": "string"
              },
              "sv-SE": {
                "type": "string"
              }
            }
          },
          "parent": {
            "$ref": "#/components/schemas/Attribute"
          },
          "parentId": {
            "type": "integer"
          }
        }
      },
      "ProductAttribute": {
        "type": "object",
        "required": ["id", "name", "value"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "value": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "productId": {
            "type": "integer"
          },
          "attribute": {
            "$ref": "#/components/schemas/Attribute"
          },
          "attributeId": {
            "type": "integer"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "object"
          }
        }
      },
...
      "Party": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "vat": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "streetNumber": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "Receipt": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "file": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "Item": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "itemNumber": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "measure": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "totalPrice": {
            "type": "number"
          },
          "totalPriceRead": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "party": {
            "$ref": "#/components/schemas/Party"
          },
          "group": {
            "$ref": "#/components/schemas/Group"
          },
          "receipts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Receipt"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
...
    }
  }
}